health and timeout as separate modules
