+++ ide/services/ide-core/pom.xml
@@ -38,6 +38,11 @@
         </dependency>
         <dependency>
 	    	<groupId>org.eclipse.dirigible</groupId>
+			<artifactId>dirigible-commons-health</artifactId>
+			<version>6.0.0-SNAPSHOT</version>
+    	</dependency>
+        <dependency>
+	    	<groupId>org.eclipse.dirigible</groupId>
 			<artifactId>dirigible-core-scheduler</artifactId>
 			<version>6.0.0-SNAPSHOT</version>
     	</dependency>
