+++ modules/commons/commons-timeout/.gitignore
@@ -1,0 +1,15 @@
+# folders
+dist/
+.settings/
+target/
+derby/
+dirigible_local/
+
+# files
+.DS_Store
+.project
+.classpath
+*.bak
+*.class
+*.jar
+derby.log
