+++ modules/commons/commons-timeout/src/main/java/org/eclipse/dirigible/commons/timeout/TimeLimited.java
@@ -9,7 +9,7 @@
  * SPDX-FileCopyrightText: 2021 SAP SE or an SAP affiliate company and Eclipse Dirigible contributors
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.dirigible.commons.config.timeout;
+package org.eclipse.dirigible.commons.timeout;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
