+++ modules/core/core-scheduler/pom.xml
@@ -37,6 +37,16 @@
 			<version>6.0.0-SNAPSHOT</version>
     	</dependency>
     	<dependency>
+	    	<groupId>org.eclipse.dirigible</groupId>
+			<artifactId>dirigible-commons-health</artifactId>
+			<version>6.0.0-SNAPSHOT</version>
+    	</dependency>
+    	<dependency>
+	    	<groupId>org.eclipse.dirigible</groupId>
+			<artifactId>dirigible-commons-timeout</artifactId>
+			<version>6.0.0-SNAPSHOT</version>
+    	</dependency>
+    	<dependency>
 		    <groupId>org.eclipse.persistence</groupId>
 		    <artifactId>jakarta.persistence</artifactId>
 		    <version>${persistence.api.version}</version>
