+++ modules/services/service-core/pom.xml
@@ -38,7 +38,7 @@
     	</dependency>
     	<dependency>
 	    	<groupId>org.eclipse.dirigible</groupId>
-			<artifactId>dirigible-commons-config</artifactId>
+			<artifactId>dirigible-commons-health</artifactId>
 			<version>6.0.0-SNAPSHOT</version>
     	</dependency>
     	<dependency>
